[Bugfix] Load Flux2 text_encoder/VAE in od_config.dtype
Mistral3ForConditionalGeneration.from_pretrained and
AutoencoderKLFlux2.from_pretrained were called without torch_dtype,
and the text_encoder config.json has torch_dtype=None. On
transformers==4.57.6 the loader then upcasts the on-disk bf16 weights
to fp32, doubling the footprint (~89.8 GiB vs ~44.9 GiB for the Mistral3
encoder alone). This makes the non-offload path OOM anything smaller
than an A100-80, and even the offload path OOMs when the encoder swaps
onto a ≤48 GiB card at text-encode time.

Pass torch_dtype=od_config.dtype (defaults to bf16) so the encoder and
VAE load in the configured precision and match the transformer.

Probe results on 1× B200 (tools/flux2_cpu_offload_probe.py):
  enable_cpu_offload=True:  0 MiB GPU at __init__ (unchanged)
  enable_cpu_offload=False: 45967.7 MiB vs 91921.8 MiB previously (~2× drop)

+++ vllm_omni/diffusion/models/flux2/pipeline_flux2.py
@@ -373,14 +373,20 @@ class Flux2Pipeline(nn.Module, CFGParallelMixin, SupportImageInput, ProgressBarM
             model, subfolder="scheduler", local_files_only=local_files_only
         )
         self.text_encoder = Mistral3ForConditionalGeneration.from_pretrained(
-            model, subfolder="text_encoder", local_files_only=local_files_only
+            model,
+            subfolder="text_encoder",
+            local_files_only=local_files_only,
+            torch_dtype=od_config.dtype,
         ).to(placement_device)
         self.tokenizer = PixtralProcessor.from_pretrained(
             model, subfolder="tokenizer", local_files_only=local_files_only
         )
-        self.vae = AutoencoderKLFlux2.from_pretrained(model, subfolder="vae", local_files_only=local_files_only).to(
-            placement_device
-        )
+        self.vae = AutoencoderKLFlux2.from_pretrained(
+            model,
+            subfolder="vae",
+            local_files_only=local_files_only,
+            torch_dtype=od_config.dtype,
+        ).to(placement_device)
         transformer_kwargs = get_transformer_config_kwargs(od_config.tf_model_config, Flux2Transformer2DModel)
         self.transformer = Flux2Transformer2DModel(**transformer_kwargs)
